Move transform functions into separate modules
